button fixes
